Cloak: now captures file list before cloaking each folder
This is the beginning of giving Cloak the ability to respond to requests by other processes for a list of files contained within a cloaked folder (directory).  This is needed because FindFirstFile and FindNextFile fail with SHARING_VIOLATION for a cloaked directory (which is the whole point of cloaking).  Cloak's API will be secured to detect usage by unauthorized client processes and will be designed to make such access difficult.