Headers refreshed
The headers used during C++ compiles are generated.  Consider the addition of a new symbolic constant.  During development, that definition will be added to a source file.  (Tyipically, it will get its own small source file.)  All of the "authoritative" source is in the "dictionary" folder.  Provisionally, the new definition will also be manually added to file "ipdos.h" in the "snip" folder.  Then the "hc" batch file is run.  Periodically, as in the instant commit, all of the generated header files are refreshed, which, for example, overwrites the ipdos.h file.  The instant commit is due to that being done.  To rebuild all of the generated headers, a particular line is uncommented within adam 621.  During the rebuild, that line will become commented out.