Regular expressions: sc_ccSSSrEPEAT can now take a special value that means "ignore"
The [idAdam,cNote] strokeS parameters normally specify the minimum and maximum number of repeats for the next search subexpression.  Now, when idAdam is -1, cNote is interpreted as the number of subexpressions to ignore.

This means that code can easily implement "closure optimization", meaning that each possible value within the repeat range can be tried, with application code then deciding which one it wants to use.  It does this by patching alternative values into the "repeat stroke".  With this new "ignore" special value, it can try ignoring the subexpression altogether.  (If N repeats are ordered, N+1 total consecutive instances of the subexpression are required.)

The motive for this enhancement is to be able to easily implement "zero or more of the following" pattern searches.