Rimstar Editor: winhlp32.exe for Windows 10
The Rimstar Programmers Editor was, and continues to be, used to craft IPDOS (tm).  That 25 year old editor is preserved within the source tree, but can no longer display its help menu because Microsoft removed support for *.hlp files from Windows versions after XP.  This commit preserves winhlp32.msi so that support for *.hlp files can be installed.  After installing that, you can open the help files that install with Rimstar.

I am a licensed user of Rimstar., which is a proprietary software product.  But I have not been able to find out its current ownership status.  Rimstar is preserved within the source tree because it is an incredibly powerful editor that outperforms anything that is available today AFAIK.  Anyone who can establish ownership is invited to contact me.  Until that happens, I preserve and disclose it on the presumption that any proprietary interests in Rimstar have been abandoned.