strRegExpFromWildcardF: maps strings containing '?' and '*' to IPDOS (tm) regular expression strokeS strings
Appears to work, but for search patterns such as "*.batonC.*", the searching takes too long.  This is because strIdOptimizedClosureF is stupid; it tries every possible length for each wildcard.

The design goal is to reproduce exactly the behavior of Windows command prompt file name searches using patterns that contain asterisks, question marks, and dots.