strAndExpressionLengthMinMaxF
This function helps closure optimization by calculating characteristics of an AND subexpression.  This information can be used to quickly eliminate closure lengths that cannot result in matches.  The information returned is:

(1) The minimum and maximum length that can match the AND expression.
(2) An [idf,idl] pair for each fixed length substring in the expression.