patternC: new overload of etherC::strIdOptimizedClosureF:
Now :strIdOptimizedClosureF can be easily called, specifying a Microsoft Windows wildcard expression, and the calling code does not need to administer the IPDOS (tm) pattern expression or its associated override parameters.

This finishes a craftwork thrust.  The objective of the thrust was to prepare to edit all code to avoid calling FindFirstFile FindNextFile with wildcards in the file name.  This can now be done.  Once that is done, diskFindFileOrDirF can be migrated to take a folderP parameter rather than a tLikeP parameter.  Once that is done, diskFindFileOrDirF can be enhanced to connect to Cloak to find out whether the requested folder is cloaked and, of so, to obtain its contents from Cloak rather than from the base o.s.

But craftwork will be suspended as of this commit so that headers will be  refreshed and then regression bugs eliminated.  (Bundle has been broken.)