header refresh
To get this header refresh to work, I had to manually delete generated source code for the SOIL (tm) subsystem.  Conjecture: this is due to the use of a patternC* parameter.  The SOIL (tm) subsystem code generator needs to be edited to support patternC* parameters.  This probably involves editing the soulC class to add support for << and >> operators for patternC.