Don't tell CreateProcess to inherit handles to the child
This interferes with Cloak if Cloak hires a child process.  I don't remember why I changed 0 to 1 about  one month ago, but I'm pretty sure that I just thought that having handles inherited would be a generally good default behavior.  I didn't have a particular need to make the change, and now I see that it is definitely NOT a good default behavior.