Cloak: design error fixed
Now Cloak recloaks the target folder before sending any info to client.  This is needed because otherwise there is a race failure:  If the client opens a handle to the folder while the folder is being recloaked,  Cloak will fail (and in production will shut down the computer).  With this change, the race is eliminated because the time of the recloaking is not associated with any event observable by the client.