tinArgS: type checking to detect collision
tinArgS is used to pass parameters recursively, to avoid C++ parameter passing overhead.  It is currently used only for listC and related code.  The code is now ready to allow tinArgS to also be used for strIdOptimizeClosureF.  This commit contains code that detects when the two different uses are attempted on the same thread at the same time.  (This cannot be done because tinArgS contains unions.)