Generated header refresh
The main generated header is "ipdos.h".  This is the complete header for the base dll.  But it is not used directly.  Instead, it in turn is used to generate the 4 header files that you see in this commit, each of which is spit into two pieces to deal with a size limit that I can no longer remember the details of.