Cool, it works!: Scratch adam 4 serves https
This is a unit test.  The certificate and private key paths are hardcoded and in the wrong place in code.  SSL_CTX context structure is constructed for each connection, which is a design error.  Code needs to be reviewed for good naming and for correct compartmentalization of functionality.  But it works!  Yippeee!