Design error workaround: http service: do not validate when opening file
The design error is that thirdC::dosOpenIF and some other image functions assume that a particular thirdC or etherC instance is to be used when calling a nonimage function.  When such a called function becomes impotent, this causes the wrong poopC instance to be used.  The effect can be to cause "smelly poop" (which terminates the process), when the caller wants the poopC within, for example, *tinP.pEtScratch, to be used, so that failure does not cause process termination.