booksC: migrate to use heapC not poolC
This migration is completed by flTHREADmODE3_POOLaLLOCATEfROMhEAP.  The next step is to reorder grab layers to put booksC at the top, just below heap.  (Conj: this will work or can be easily made to work.)  This is a "Gordian Knot" cutting attempt.  If it doesn't work, I will edit booksC to eliminate all code that uses poolC.