milestone: inOutFrameC audit:  cpu cycle app telemetry is audited
Every ~inOutFrameC for which telemetry capture is enabled verifies that cpu cycle audit accumulators updated by its kids total to the same cpu cycle value that it calculates between the end of its own ct and the beginning of its own dt.  Capture is enabled practically everywhere, with sole exceptions being when app telemetry is being flushed to book or when other code writes to app telemetry book.