milestone: app telemetry finished but won't work (design problem)
The code design erroneously ct's processGlobal3S::bks* specifying "do not serialize".  This appears to be due to the fact that the grabWantS array napkinC, which is required for grabC, is not ct until later.  But a quick investigation suggests that ct'ing the array napkinC earlier is feasible, and would allow the bks members to be serialized, which is required for correct function.