bookC: flBOOKc_USEhEAPnOTnAPKIN
This new feature supports ct of bookC instance using heap rather than a napkinC.  Since no system telemetry is generated, such bookC instances can be ct earlier and dt later, i.e. anywhere within the life of processGlobal3I.heap .  Motivation: Expectation that such instances can be used to permit early queueing of sys and app telemetry for deferred output, which will in turn allow ct of the grabWantS array to be ct earlier, which will in turn allow ct of the telemetry and other books in processGlobal3S to be ct with serialization, which is needed.  (It was a design error to assume that such serialization is not needed; it is.)