work in progress: before moving tmcHttpServerF code to etherC member function
This will be so that ifcDrivenServerF can OPTIONALLY use a worker thread pool.  If no, then TASKdRIVENsERVER will be used, as in current code.  If yes, then TASK will be used, and an ooth thread main function will be needed that does the TASKdRIVENsERVER function for each job that it pulls from stJobsP.