milestone: http cached when no query
sexC::newF sometimes detects duplicate "untagged" registration.  This can be studied by tracing new and del of pbo to see whether the same pbo is returned oo times without any deletion.  That is what is implied, that the same pbo is returned to the offending thread 2 times without an intervening deletion ON THAT THREAD.  Note that this code will fail if another thread deletes the pbo.