http: badge: proof of valid credentials
httpServerC now dispenses "badges", each of which is a random countT value associated with an nnPeer value that proves that the nnPeer provided valid credentials.  The intent here is to permit etherDoHttpJobF to serve the correct flavor (logged in or not) of a page without needing to validate credentials each time, so that the page can be served from cache.  Issue: when there are multiple servers, the list of [badge,nnPeer] pairs needs to be coherent across all servers.  This commit only assigns a badge value locally, and does not use it or broadcast it to other servers.