readRequestF: bug fix to eliminate memory corruption exploit
If an http request is received that declares a false, understated content length, and the portion of the request body received with the header is longer, pool memory would be corrupted because the declared content length was used to allocate the receiving buffer.  With this commit, the allocation is increased to enable the content to be received without overwriting memory.  This behavior is almost certainly an intentional attack.