sexC slotsC poolC and others: rename variables for readability
Use variable and constant names to consistently distinguish between a "drop" and a "slot".  A pool "drop" is the portion of a "slot" that is available for use by the application code that calls poolC::newF.  A "slot" is the unit of allocation internally to sexC.  The layout of each slot is [fingerSeparator,dropHeader,drop].

This commit is only  code renaming commit.  There should be no semantic (execution logic) changes within this commit.