example simplest: marriage model encryption over socketC
This commit fixes the code implementing marriage model encryption, which is a "homenade" symmetric encryption method that essentially uses the entire history of communication between oo peers as the shared secret.  It is secure only to the extent that an eavesdropper can be prevented from observing the entire history.  In that situation, the strength of the method is unknown.