milestone: dut works (relayC channel i/o factory1 factory4
The bug was in idPortC, the == operator functions.  They were only looking at idPort, not also at nn.  This suggests that when I was working with relayC before, I was only using multiple homes on a single computer, so never realized that this bug was there.  The current dut tests wo home on each of oo computers, and can easily be extended to test oo homes on each of ni computers.