c_getaddrinfoF
Craftwork in progress. c_getaddrinfoF successfully gets 2 addresses, including the IPv4 (verified) and presumably the IPv6 (not verified) address of factory6.ideafarm.com .  This code was difficult apparently because of a header error in the win32 headers packaged with Open Watcom.  Now that it works, the next step is to finish developing c_getaddrinfoF as a thin wrapper, and then use that new api to obtain all (including IPv6) addresses for a name.