Eliminate hardcoded development drive letter.
With this commit, build can be done on any drive, and nothing needs to be edited before building.  The only thing needed is to enable .NET 3.5 (which is required for Wix Toolset.  The build zip file contains OW and Wix and will use its copies if OW and/or Wix toolset is not installed.