adam a: baseline for rewrite of relayOldC
Adam a runs channel i/o using its own copy of all of the relayC subsystem code.  My plan is to incrementally rewrite the entire relayC class, using the code that is now in adam a as a baseline.  Adam a currently runs aok; it duplicates the functionality of adam 115 (kernel3bang).