unit test pass: nodeC (lan): maintains stOpen (but incorrectly)
dut: factory1 factory4 each running watch and adam b
when factory4 is unplugged from switch, both cells of adam b detect that the other factory is no longer open.  When factory4 is plugged back in, both cells of adam b detect this and add the other factory to stOpen.  But each adam b cell attributes the failure to the other factory, so the ordering no longer matches due to the timeS field.  (the faulting factory gets a new timestamp, which puts it lath in stOpen.)

This needs to be fixed.  Unplugging a wire and restoring it should not cause the stOpen list to go out of sync.