header refresh ; nodeC development milestone
The new nodeC code (in adam b) is now organized so that it can easily be enhanced to support global mode.  Mode affects only the allocation and populating of pNodeList.  For local mode, pNodeList points to the napkinC maintained by adam WATCH.  For global mode, pNodeList will point to a poolC drop that is populated to list nodes registered in DNS as well as global nodes remembered from prior executions.