move registry container to new Rackspace account
The code contains credentials for oo Rackspace accounts.  Credentials that allow writing are in code so that code can push its registration file, which is processed by factory1.  The registry container is now in the new account that only contains the registry container.  The other credentials are read only and are used to read the privileges container.