debugging in progress
The dut is still at will sameo immediately.  tlsCheckF and other tls code contains debugging code to try to detect data inconsistency.  This is a puzzle because the interaction of multiple threads appears to be needed for sameo, but all of this code should be operating on data visible only to the current thread.