debugging hunt in progress
The dut is adam c, which fails in tlsCheckF within a tock.  Often, the call to tlsCheckF is by ~inOutFrameC, which is in turn called by dosSuspendThreadIF, which does practically nothing.  All of the fails are consistent with the hypothesis that the memory inspected by tlsCheckF is being modified by another thread rather than by the calling thread.