nodeC: now has tmLiveF
This thread makes each node "alive" in that it can initiate action.  tmLiveF walks stOpenP and then sleeps for a tock before doing it again.  It knows whether it is the root node (plate 1 in stOpen).  The next step is to have each node that thinks that it is the root node emit an "i am root" command.