exception reporting:  exception reporting working for main thread
When a process is not a child or the kernel monitor, an exception on any thread will produce an exception report file.  Now, with this commit, if such a process is a child of the kernel monitor, the same report file will be produced, and in addition the kernel monitor will report the event in an exception minute report.

(This commit adds a _try _except block set to the main thread to accomplish this, with a filter function that calls the same handler that  the process registers as the default top level handler.)