_try _except for kid threads
With this commit, all kid threads defined using TASK and DONE are executed within _try _except so that an exception file will be produced even if the process is run as a child of kernel monitor (which "debugs" its child processes).

Regression testing for restartC functionality is needed and has not yet been done.